+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Latest changes
 
+* VFS-13155 Fixed display of longer membership paths when the display limit is reached
 * VFS-13178 Fixed privileges bulk edit to support selecting groups from multiple pages
 * VFS-13221 Added support for new Calendar Versioning scheme in Onedata (e.g. version 25.0)
 * VFS-13213 Fixed blank Oneprovider domain in providers map popover